#BTR04:refactor/Alterando de acordo com o diagrama e arrumando herança

+++ src/main/java/com/liratech/helppsico/infrastructure/repositories/entities/AvaliacaoEntity.java
@@ -26,6 +26,6 @@ public class AvaliacaoEntity {
     @JoinColumn(name = "id_paciente")
     private PacienteEntity paciente;
 
-    public BigDecimal nota;
+    public Double nota;
     public String comentario;
 }
